// packages/test-ignore_export_console/gen/ignore-export-console.ts
export interface ConsolePrivatePerson {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * @generated from protobuf field: int32 age = 2
     */
    age: number;
    /**
     * @generated from protobuf field: repeated ConsoleAssetInfo assets = 3
     */
    assets: ConsoleAssetInfo[];
}
/**
 * @generated from protobuf message ConsoleReq
 */
export interface ConsoleReq {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string;
}
/**
 * @generated from protobuf message ConsoleResp
 */
export interface ConsoleResp {
    /**
     * @generated from protobuf field: string message = 1
     */
    message: string;
}
/**
 * @generated from protobuf message GetPrivatePersonConsoleInfoReq
 */
export interface GetPrivatePersonConsoleInfoReq {
    /**
     * @generated from protobuf field: string personId = 1
     */
    personId: string;
}
/**
 * @generated from protobuf message GetPrivatePersonConsoleInfoResp
 */
export interface GetPrivatePersonConsoleInfoResp {
    /**
     * @generated from protobuf field: ConsolePrivatePerson privatePerson = 1
     */
    privatePerson?: ConsolePrivatePerson;
}
// ...
export interface GetEmbeddedConsoleMessageWithEnumReq {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string;
    /**
     * @generated from protobuf field: GetEmbeddedConsoleMessageWithEnumReq.Day day = 2
     */
    day: GetEmbeddedConsoleMessageWithEnumReq_Day;
    /**
     * @generated from protobuf field: GetEmbeddedConsoleMessageWithEnumReq.WeatherInfo weatherInfo = 3
     */
    weatherInfo?: GetEmbeddedConsoleMessageWithEnumReq_WeatherInfo;
    /**
     * @generated from protobuf field: ConsoleAddress address = 4
     */
    address?: ConsoleAddress;
}
// ...
export enum GetEmbeddedConsoleMessageWithEnumReq_Day {
    /**
     * @generated from protobuf enum value: UNKNOWN_DAY = 0;
     */
    UNKNOWN_DAY = 0,
    /**
     * @generated from protobuf enum value: MONDAY = 1;
     */
    MONDAY = 1,
    /**
     * @generated from protobuf enum value: TUESDAY = 2;
     */
    TUESDAY = 2,
    /**
     * @generated from protobuf enum value: WEDNESDAY = 3;
     */
    WEDNESDAY = 3,
    /**
     * @generated from protobuf enum value: THURSDAY = 4;
     */
    THURSDAY = 4,
    /**
     * @generated from protobuf enum value: FRIDAY = 5;
     */
    FRIDAY = 5,
    /**
     * @generated from protobuf enum value: SATURDAY = 6;
     */
    SATURDAY = 6,
    /**
     * @generated from protobuf enum value: SUNDAY = 7;
     */
    SUNDAY = 7
}
/**
 * @generated from protobuf message GetEmbeddedConsoleMessageWithEnumResp
 */
export interface GetEmbeddedConsoleMessageWithEnumResp {
    /**
     * @generated from protobuf field: ConsoleAddress address = 1
     */
    address?: ConsoleAddress;
    /**
     * @generated from protobuf field: GetEmbeddedConsoleMessageWithEnumResp.Msg msg = 2
     */
    msg?: GetEmbeddedConsoleMessageWithEnumResp_Msg;
}
// ...
export const ConsolePrivatePerson = new ConsolePrivatePerson$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ConsoleReq$Type extends MessageType<ConsoleReq> {
    constructor() {
        super("ConsoleReq", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ConsoleReq>): ConsoleReq {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        if (value !== undefined)
            reflectionMergePartial<ConsoleReq>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ConsoleReq): ConsoleReq {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ConsoleReq, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ConsoleReq
 */
export const ConsoleReq = new ConsoleReq$Type();
// @generated message type with reflection information, may provide speed optimized methods
class ConsoleResp$Type extends MessageType<ConsoleResp> {
    constructor() {
        super("ConsoleResp", [
            { no: 1, name: "message", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<ConsoleResp>): ConsoleResp {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.message = "";
        if (value !== undefined)
            reflectionMergePartial<ConsoleResp>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ConsoleResp): ConsoleResp {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string message */ 1:
                    message.message = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ConsoleResp, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string message = 1; */
        if (message.message !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.message);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ConsoleResp
 */
export const ConsoleResp = new ConsoleResp$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetPrivatePersonConsoleInfoReq$Type extends MessageType<GetPrivatePersonConsoleInfoReq> {
    constructor() {
        super("GetPrivatePersonConsoleInfoReq", [
            { no: 1, name: "personId", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<GetPrivatePersonConsoleInfoReq>): GetPrivatePersonConsoleInfoReq {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.personId = "";
        if (value !== undefined)
            reflectionMergePartial<GetPrivatePersonConsoleInfoReq>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetPrivatePersonConsoleInfoReq): GetPrivatePersonConsoleInfoReq {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string personId */ 1:
                    message.personId = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetPrivatePersonConsoleInfoReq, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string personId = 1; */
        if (message.personId !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.personId);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message GetPrivatePersonConsoleInfoReq
 */
export const GetPrivatePersonConsoleInfoReq = new GetPrivatePersonConsoleInfoReq$Type();
// @generated message type with reflection information, may provide speed optimized methods
class GetPrivatePersonConsoleInfoResp$Type extends MessageType<GetPrivatePersonConsoleInfoResp> {
    constructor() {
        super("GetPrivatePersonConsoleInfoResp", [
            { no: 1, name: "privatePerson", kind: "message", T: () => ConsolePrivatePerson }
        ]);
    }
    create(value?: PartialMessage<GetPrivatePersonConsoleInfoResp>): GetPrivatePersonConsoleInfoResp {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<GetPrivatePersonConsoleInfoResp>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: GetPrivatePersonConsoleInfoResp): GetPrivatePersonConsoleInfoResp {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* ConsolePrivatePerson privatePerson */ 1:
                    message.privatePerson = ConsolePrivatePerson.internalBinaryRead(reader, reader.uint32(), options, message.privatePerson);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: GetPrivatePersonConsoleInfoResp, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* ConsolePrivatePerson privatePerson = 1; */
        if (message.privatePerson)
            ConsolePrivatePerson.internalBinaryWrite(message.privatePerson, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message GetPrivatePersonConsoleInfoResp
 */
export const GetPrivatePersonConsoleInfoResp = new GetPrivatePersonConsoleInfoResp$Type();
// ...
export const GetEmbeddedConsoleMessageWithEnumReq = new GetEmbeddedConsoleMessageWithEnumReq$Type();
// ...
export const GetEmbeddedConsoleMessageWithEnumResp = new GetEmbeddedConsoleMessageWithEnumResp$Type();
// ...
export const IgnoreExportConsoleService = new ServiceType("IgnoreExportConsoleService", [
    { name: "EmptyOptionConsoleMethod", options: {}, I: ConsoleReq, O: ConsoleResp },
    { name: "MethodWithExportConsole1", options: { "blocker.ignoreExportConsole": 0 }, I: ConsoleReq, O: ConsoleResp },
    { name: "GetPrivatePersonConsoleInfo", options: { "blocker.ignoreExportConsole": 0 }, I: GetPrivatePersonConsoleInfoReq, O: GetPrivatePersonConsoleInfoResp },
    { name: "GetEmbeddedConsoleMessageWithEnum", options: { "blocker.ignoreExportConsole": 0 }, I: GetEmbeddedConsoleMessageWithEnumReq, O: GetEmbeddedConsoleMessageWithEnumResp }
]);